+++ webknossos/webknossos/dataset/dataset.py
@@ -33,10 +33,7 @@ from upath import UPath
 
 from ..geometry.vec3_int import Vec3Int, Vec3IntLike
 from ._array import ArrayException, ArrayInfo, BaseArray, DataFormat
-from .remote_dataset_registry import (
-    RemoteDatasetRegistry,
-    RemoteOrganizationDatasetRegistry,
-)
+from .remote_dataset_registry import RemoteDatasetRegistry
 
 if TYPE_CHECKING:
     import pims
@@ -1367,29 +1364,25 @@ class Dataset:
             )
 
     @staticmethod
-    def get_remote_dataset_registry() -> Mapping[
-        str, RemoteOrganizationDatasetRegistry
-    ]:
+    def get_remote_datasets(
+        organization_id: Optional[str] = None,
+        tags: Optional[Union[str, Sequence[str]]] = None,
+    ) -> Mapping[str, "RemoteDataset"]:
         """
-        Returns a registry of all remote datasets visible for the current user. The registry is built as a dict of dicts.
-        The outer dict-keys are the organization ids of the datasets, the inner keys are the dataset names:
+        Returns a dict of all remote datasets visible for selected organization, or the organization of the logged in user by default.
+        The dict contains lazy-initialized `RemoteDataset` values for keys indicating the dataset name.
+
         ```python
         import webknossos as wk
 
-        dataset_registry = wk.Dataset.get_remote_dataset_registry()
-        dataset_registry["scalable_minds"]["l4dense_motta_et_al_demo"]
+        print(sorted(wk.Dataset.get_remote_datasets()))
 
-        current_organization = wk.User.get_current_user().organization_id
-        sorted(dataset_registry[current_organization])
-        ```
-
-        After selecting an organization, datasets can also be accessed by tag or by their display name:
-        ```python
-        dataset_registry["scalable_minds"].by_tag["demo"]
-        dataset_registry["scalable_minds"].by_display_name["L4 Mouse Cortex Demo"]
+        ds = wk.Dataset.get_remote_datasets(
+            organization_id="scalable_minds"
+        )["l4dense_motta_et_al_demo"]
         ```
         """
-        return RemoteDatasetRegistry()
+        return RemoteDatasetRegistry(organization_id=organization_id, tags=tags)
 
 
 class RemoteDataset(Dataset):