+++ webknossos/webknossos/dataset/remote_dataset_registry.py
@@ -1,7 +1,4 @@
-from collections import defaultdict
-from typing import TYPE_CHECKING, DefaultDict, Dict, List, Mapping, TypeVar
-
-from boltons.dictutils import FrozenDict
+from typing import TYPE_CHECKING, Optional, Sequence, TypeVar, Union
 
 from webknossos.utils import LazyReadOnlyDict
 
@@ -14,74 +11,38 @@ V = TypeVar("V")  # value
 C = TypeVar("C")  # cache
 
 
-class RemoteOrganizationDatasetRegistry(LazyReadOnlyDict[str, "RemoteDataset"]):
+class RemoteDatasetRegistry(LazyReadOnlyDict[str, "RemoteDataset"]):
     """Dict-like class mapping dataset names to `RemoteDataset` instances."""
 
-    by_display_name: Mapping[str, "RemoteDataset"]
-    by_tag: Mapping[str, List["RemoteDataset"]]
-
     def __init__(
         self,
-        organization_id: str,
-        names: List[str],
-        by_display_name: Dict[str, str],
-        by_tag: Dict[str, List[str]],
+        organization_id: Optional[str],
+        tags: Optional[Union[str, Sequence[str]]],
     ) -> None:
-        from webknossos.dataset.dataset import Dataset
-
-        super().__init__(
-            entries=dict(zip(names, names)),
-            func=lambda name: Dataset.open_remote(name, organization_id),
-        )
-        self.by_display_name = LazyReadOnlyDict(
-            entries=by_display_name,
-            func=lambda name: Dataset.open_remote(name, organization_id),
-        )
-        self.by_tag = LazyReadOnlyDict(
-            entries=by_tag,
-            func=lambda names: [
-                Dataset.open_remote(name, organization_id) for name in names
-            ],
-        )
-
-
-class RemoteDatasetRegistry(
-    FrozenDict, Mapping[str, RemoteOrganizationDatasetRegistry]
-):
-    """Dict-like class mapping organization ids to `RemoteOrganizationDatasetRegistry` instances."""
-
-    def __init__(
-        self,
-    ) -> None:
+        from webknossos.administration.user import User
         from webknossos.client._generated.api.default import dataset_list
         from webknossos.client.context import _get_generated_client
+        from webknossos.dataset.dataset import Dataset
 
         client = _get_generated_client(enforce_auth=True)
+
+        if organization_id is None:
+            organization_id = User.get_current_user().organization_id
+
+        if isinstance(tags, str):
+            tags = [tags]
+
         response = dataset_list.sync(client=client)
         assert response is not None
 
-        names_per_org: DefaultDict[str, List[str]] = defaultdict(list)
-        by_display_name_per_org: DefaultDict[str, Dict[str, str]] = defaultdict(dict)
-        by_tag_per_org: DefaultDict[str, DefaultDict[str, List[str]]] = defaultdict(lambda: defaultdict(list))  # type: ignore[call-overload]
+        ds_names = []
 
         for ds_info in response:
-            names_per_org[ds_info.owning_organization].append(ds_info.name)
-            if ds_info.display_name:
-                by_display_name_per_org[ds_info.owning_organization][
-                    ds_info.display_name
-                ] = ds_info.name
-            for tag in ds_info.tags:
-                by_tag_per_org[ds_info.owning_organization][tag].append(ds_info.name)
+            tags_match = tags is None or any(tag in tags for tag in ds_info.tags)
+            if tags_match:
+                ds_names.append(ds_info.name)
 
         super().__init__(
-            (
-                org,
-                RemoteOrganizationDatasetRegistry(
-                    org,
-                    names=names_per_org[org],
-                    by_display_name=by_display_name_per_org[org],
-                    by_tag=by_tag_per_org[org],
-                ),
-            )
-            for org in names_per_org
+            entries=dict(zip(ds_names, ds_names)),
+            func=lambda name: Dataset.open_remote(name, organization_id),
         )